Fetch vulnerabilities post Product imports #94
