+++ product_portfolio/models.py
@@ -29,6 +29,7 @@ from component_catalog.models import KeywordsMixin
 from component_catalog.models import LicenseExpressionMixin
 from component_catalog.models import Package
 from component_catalog.models import component_mixin_factory
+from component_catalog.vulnerabilities import fetch_for_queryset
 from dje import tasks
 from dje.fields import LastModifiedByField
 from dje.models import DataspacedManager
@@ -489,6 +490,10 @@ class Product(BaseProductMixin, FieldChangesMixin, KeywordsMixin, DataspacedMode
             user_uuid=user.uuid,
             dataspace_uuid=user.dataspace.uuid,
         )
+
+    def update_vulnerabilities(self):
+        """Update the vulnerabilties of all the Package realted to this Product."""
+        return fetch_for_queryset(self.all_packages, self.dataspace)
 
 
 class ProductRelationStatus(BaseStatusMixin, DataspacedModel):
